pyproject.toml
[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vitaclaw"
version = "2.0.0"
description = "Modular health AI skill library for OpenClaw"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
viz = [
    "matplotlib>=3.8",
    "pandas>=2.2.3",
    "scipy>=1.11",
]
ocr = [
    "paddleocr>=3.4.0",
    "paddlepaddle>=3.0",
]
oncology = [
    "numpy>=1.26",
    "pandas>=2.2.3",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.9",
]

[tool.setuptools.packages.find]
include = ["skills*"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
